+++ api/models.py
@@ -1,4 +1,6 @@
 import json
+import StringIO
+import logging
 import os
 from datetime import datetime
 import cPickle as pickle
@@ -94,16 +96,17 @@ class ImportHandler(Document):
     use_dot_notation = True
 
     def create_dataset(self, params, run_import_data=True):
-        from api.utils import slugify
+        #from api.utils import slugify
         dataset = app.db.DataSet()
         str_params = "-".join(["%s=%s" % item
                               for item in params.iteritems()])
         dataset.name = "%s: %s" % (self.name, str_params)
         dataset.import_handler_id = str(self._id)
         dataset.import_params = params
-        filename = '%s-%s.json' % (slugify(self.name), str_params.replace('=', '_'))
-        dataset.data = filename
+        # filename = '%s-%s.json' % (slugify(self.name), str_params.replace('=', '_'))
+        # dataset.data = filename
         dataset.save(validate=True)
+        dataset.set_file_path()
         return dataset
 
     def delete(self):
@@ -141,6 +144,8 @@ class Model(Document):
     """
     STATUS_NEW = 'New'
     STATUS_QUEUED = 'Queued'
+    STATUS_IMPORTINING = 'Importing'
+    STATUS_IMPORTED = 'Imported'
     STATUS_TRAINING = 'Training'
     STATUS_TRAINED = 'Trained'
     STATUS_ERROR = 'Error'
@@ -202,12 +207,12 @@ class Model(Document):
 
     def run_test(self, dataset, callback=None):
         trainer = self.get_trainer()
-        path = app.config['DATA_FOLDER']
-        if not exists(path):
-            makedirs(path)
-        with open(join(path, dataset.data), 'r') as train_fp:
-            metrics = trainer.test(streamingiterload(train_fp), callback=callback)
-        train_fp.closed
+        fp = dataset.get_data_stream()
+        try:
+            metrics = trainer.test(streamingiterload(fp), callback=callback)
+        finally:
+            fp.close()
+
         raw_data = trainer._raw_data
         trainer.clear_temp_data()
         return metrics, raw_data
@@ -225,6 +230,12 @@ class Model(Document):
         self.delete_metadata()
         self.collection.remove({'_id': self._id})
 
+    def set_error(self, error, commit=True):
+        self.error = str(error)
+        self.status = self.STATUS_ERROR
+        if commit:
+            self.save()
+
     def delete_metadata(self):
         params = {'model_id': str(self._id)}
         app.db.Test.collection.remove(params)
@@ -236,6 +247,8 @@ class Model(Document):
 @connection.register
 class Test(Document):
     STATUS_QUEUED = 'Queued'
+    STATUS_IMPORTINING = 'Importing'
+    STATUS_IMPORTED = 'Imported'
     STATUS_IN_PROGRESS = 'In Progress'
     STATUS_COMPLETED = 'Completed'
     STATUS_ERROR = 'Error'
@@ -326,13 +339,64 @@ class DataSet(Document):
         'data': basestring,
         'import_params': dict,
         'import_handler_id': basestring,
+        'on_s3': bool,
+        'compress': bool,
+        'filename': basestring,
     }
     required_fields = ['name', 'created_on', 'updated_on', ]
     default_values = {'created_on': datetime.utcnow,
                       'updated_on': datetime.utcnow,
                       'error': '',
+                      'on_s3': False,
+                      'compress': True,
                       'status': STATUS_IMPORTINING}
     use_dot_notation = True
+
+    def __init__(self, *args, **kwargs):
+        super(DataSet, self).__init__(*args, **kwargs)
+
+    def set_file_path(self):
+        self.data = '%s.%s' % (self._id, 'gz' if self.compress else 'json')
+        path = app.config['DATA_FOLDER']
+        if not exists(path):
+            makedirs(path)
+        self.filename = join(path, self.data)
+        self.save()
+
+    def get_data_stream(self):
+        import gzip
+        #import zlib
+        if self.on_s3:
+            logging.info('Loading data from Amazon S3')
+            data = self.load_from_s3()
+            stream = StringIO.StringIO(data)
+            if self.compress:
+                logging.info('Decompress data')
+                return gzip.GzipFile(fileobj=stream, mode='r')
+                #data = zlib.decompress(data)
+            return stream
+        else:
+            logging.info('Loading data from local file')
+            open_meth = gzip.open if self.compress else open
+            return open_meth(self.filename, 'r')
+
+    def load_from_s3(self):
+        from api.amazon_utils import AmazonS3Helper
+        helper = AmazonS3Helper()
+        return helper.load_key(str(self._id))
+
+    def save_to_s3(self):
+        from api.amazon_utils import AmazonS3Helper
+        meta = {'handler': self.import_handler_id,
+                'dataset': self.name,
+                'params': str(self.import_params)}
+        helper = AmazonS3Helper()
+        helper.save_key(str(self._id), self.filename, meta)
+        #helper.save_gz_file(str(self._id), self.filename, meta)
+        helper.close()
+        #logging.info("Keys in bucket: %s" % [i for i in helper.bucket.list()])
+        self.on_s3 = True
+        self.save()
 
     def set_error(self, error, commit=True):
         self.error = str(error)
@@ -343,12 +407,14 @@ class DataSet(Document):
     def delete(self):
         super(DataSet, self).delete()
         # TODO: check import handler type
-        data_folder = app.config['DATA_FOLDER']
-        filename = os.path.join(data_folder, self.data)
         try:
-            os.remove(filename)
+            os.remove(self.filename)
         except OSError:
             pass
+        if self.on_s3:
+            from api.amazon_utils import AmazonS3Helper
+            helper = AmazonS3Helper()
+            helper.delete_key(str(self._id))
 
     def save(self, *args, **kwargs):
         if self.status != self.STATUS_ERROR: